+++ functions.php
@@ -155,6 +155,11 @@ require get_template_directory() . '/inc/template-functions.php';
 require get_template_directory() . '/inc/customizer.php';
 
 /**
+ * Pagination.
+ */
+require get_template_directory() . '/inc/pagination.php';
+
+/**
  * Load Widgets file.
  */
 require get_template_directory() . '/inc/widgets.php';