+++ index.php
@@ -16,8 +16,8 @@
 
 <?php get_header(); ?>
 
-<div id="primary" class="content-area col-md-8">
-	<main id="main" class="site-main">
+<div class = "row">
+    <div id="primary" class="content-area col-md-8">
 
 		<?php
 		if ( have_posts() ) :
@@ -42,7 +42,8 @@
 
 			endwhile;
 
-			the_posts_navigation();
+			// The pagination component -->
+			simplicityzen_pagination();
 
 		else :
 
@@ -50,11 +51,11 @@
 
 		endif; ?>
 
-	</main><!-- #main -->
-</div><!-- #primary -->
+	</div><!-- #primary -->
 
-<div class="col-md-4">
-	<?php get_sidebar(); ?>
-</div>
+	<div class="col-md-4">
+		<?php get_sidebar(); ?>
+	</div>
+</div><!-- end of row -->
 
 <?php get_footer();