+++ single.php
@@ -9,8 +9,8 @@
 
 get_header(); ?>
 
-	<div id="primary" class="content-area">
-		<main id="main" class="site-main">
+<div class = "row">
+    <div id="primary" class="content-area col-md-8">
 
 		<?php
 		while ( have_posts() ) : the_post();
@@ -27,9 +27,9 @@ get_header(); ?>
 		endwhile; // End of the loop.
 		?>
 
-		</main><!-- #main -->
 	</div><!-- #primary -->
 
-<?php
-get_sidebar();
-get_footer();
+	<div class="col-md-4">
+		<?php get_sidebar(); ?>
+	</div>
+</div><!-- end of row -->