+++ CHANGELOG.md
@@ -8,3 +8,8 @@
 - ** Release 0.6.3 03/06/2018 **
   - upgrade fontawesome version to 5.0.8
   - change fontasesome method from SVG with JS to WebFont with CSS
+
+- ** Release 0.6.4 03/08/2018 **
+  - add screenshot.png
+  - add Bootstrap 4 Pagination function
+  - update the index.php,single.php,search.php to bootstrap 4 grid layout